Reorganize hierarchy
