--- app/index.html
@@ -1,14 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-  <meta charset="UTF-8">
-  <title>Miew via React</title>
-
-  <script src="https://unpkg.com/@babel/polyfill@7/dist/polyfill.min.js"></script>
-</head>
-<body>
-  <h1>Application demonstrating Miew</h1>
-  <div class='main'>Loading Miew...</div>
-  <script src="index.js"></script>
-</body>
-</html>
