+++ app/src/components/App.scss
@@ -1,6 +1,6 @@
 /* stylelint-disable no-descending-specificity */
 
-@import "miew/dist/Miew.scss";
+@import "miew/dist/Miew.css";
 
 //////////////////////////////////////////////////////////////////////////////
 // Full-page web application
