+++ app/src/index.html
@@ -1,0 +1,10 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <title>Template React application using Miew</title>
+</head>
+<body>
+  <div id="miew-react-app">Miew wrapper as a React component.</div>
+</body>
+</html>
