+++ app/src/index.js
@@ -1,0 +1,7 @@
+import React from 'react';
+import ReactDOM from 'react-dom';
+import './components/App.scss';
+import App from './components/App';
+
+const root = document.getElementById('miew-react-app');
+ReactDOM.render(<App/>, root);
